move scheduleController to scheduler package
